Break 'attributes' into separate module
PLAT-1860
